DDD: Diversity-Dependent Diversification

Implements maximum likelihood and bootstrap methods based on the diversity-dependent birth-death process to test whether speciation or extinction are diversity-dependent, under various models including various types of key innovations. See Etienne et al. 2012, Proc. Roy. Soc. B 279: 1300-1309, doi:10.1098/rspb.2011.1439 , Etienne & Haegeman 2012, Am. Nat. 180: E75-E89, doi:10.1086/667574 , Etienne et al. 2016. Meth. Ecol. Evol. 7: 1092-1099, doi:10.1111/2041-210X.12565 and Laudanno et al. 2021. Syst. Biol. 70: 389–407, doi:10.1093/sysbio/syaa048 . Also contains functions to simulate the diversity-dependent process.

See also

Author

Maintainer: Rampal S. Etienne r.s.etienne@rug.nl (ORCID)

Authors:

Bart Haegeman bart.haegeman@sete.cnrs.fr (ORCID)

Other contributors:

Hanno Hildenbrandt h.hildenbrandt@rug.nl (ORCID) [contributor]

Giovanni Laudanno glaudanno@gmail.com (ORCID) [contributor]