Add a Service to a Domain

Domain Service

add_domain_service(name, domain = NULL)

Arguments

name
(character) Domain Service name.

domain
(character) Domain Service domain name.

Details

Don't confuse a Domain Service with an Application Service. If you are looking for a component that enables a remote client to interact with your system, then you need an Application Service, not a Domain Service.

Application Services serve clients by orchestrating Domain Services. Domain Services are not aware and have no knowledge of any Application Service.

Overview

Entities and Value Objects are often to fine-grained to provide a convenient access to the capabilities of the domain layer.

Domain Service is part of the domain model. That means it expresses Ubiquitous

Language in its Bounded Context.

Characteristics of a good Domain Service

The operation relates to a domain concept that is not a natural part of an Entity or Value Object.

The interface is defined in terms of other elements of the domain model.

The operation is stateless.

Make sure you need a service. Stripping away behaviors that belong to Entity overzealously, lead to Anemic Domain Model, where Entities comprise data but not behavior. This is similar to representing an Entity with only a `data.frame`.

Unit of Work (UoW)

There are no complex layers here. At most this component runs a query against the query store database and maybe serializes the query result into some format for transport (maybe a DTO, but maybe not), if that’s necessary. [@Vernon2013, p. 141]

UoW Abstract Base Class

Similar to python abc module

Has the following public methods:

new i.e. initialize

commit must be implemented

rollback must be implemented

exit i.e. finalize. This function which calls rollback

UoW Implementations

UoW Database Implementing

DBI, dbplyr, RSQLite

UoW Data Structure Implementing

R6DS, mainly R6DS::RDict

References

[@Vernon2013, p. 141]: R:@Vernon2013,%0Ap.%20141 [python abc module]: R:python%20abc%20module

See also