Make a Project Independent of the ddd Package

Make a Project Independent of the ddd Package

use_ddd(path = ".")

Arguments

path
(character) Path to the root folder of the project.

Details

What does use_ddd do?

Given a path of the project root folder,

When use_ddd is called

use_ddd(path = '.')

Then the function copies the following files into the path/R:

ddd-abc.R

With these files, a project can run domain objects created by ddd without having the package installed.

Workflow

Call use_ddd maximum once per project; this will make the project independent of ddd

Call devtools::document (or Ctrl+Shift+D on windows) to include the copied classes in the project NAMESPACE file